Replace Actions with namespaced actions
